+++ src/Middleware/ExceptionResponder.php
@@ -78,7 +78,10 @@ final class ExceptionResponder implements MiddlewareInterface
 
                     if (is_callable($responseHandler)) {
                         /** @var ResponseInterface */
-                        return $this->injector->invoke($responseHandler, ['exception' => $t]);
+                        return $this->injector->invoke($responseHandler, [
+                            'exception' => $t,
+                            'request' => $request,
+                        ]);
                     }
                 }
             }